Tests fix usage same file

See merge request unsreg/prj_hillel_java_ee_201912!41

+++ eugene-kovalov/src/test/java/ua/ithillel/dnepr/eugenekovalov/CityCrudRepoTest.java
@@ -9,31 +9,38 @@ import ua.ithillel.dnepr.eugenekovalov.repository.entity.City;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.ThreadLocalRandom;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class CityCrudRepoTest {
 
-    static Path pathToOrigin = Paths.get("src/main/resources/city.csv");
-    static Path pathToWorkingCopy = Paths.get("src/test/resources/tmpCity.csv");
-
-    CrudRepoImpl<City, Integer> cityIntegerCrudRepo = new CrudRepoImpl<>(pathToWorkingCopy, City.class);
+    static Path pathToWorkingCopy = null;
+    static CrudRepoImpl<City, Integer> cityIntegerCrudRepo;
 
     @SneakyThrows
     @BeforeAll
     static void prepare() {
-        clean();
-        Files.copy(pathToOrigin, pathToWorkingCopy);
+        pathToWorkingCopy = Files.createTempFile("tmpCityK", ".csv");
+        cityIntegerCrudRepo = new CrudRepoImpl<>(pathToWorkingCopy, City.class);
+
+        cityIntegerCrudRepo.create(createCity(1, 1, 1, "Dunwich"));
+        cityIntegerCrudRepo.create(createCity(2, 3, 6, "Dunwich"));
+        cityIntegerCrudRepo.create(createCity(423, 4, 7, "Darn"));
+        cityIntegerCrudRepo.create(createCity(777, 24, 71, "Maluoka"));
     }
 
     @SneakyThrows
     @AfterAll
     static void clean() {
-        Files.deleteIfExists(pathToWorkingCopy);
+        if (pathToWorkingCopy != null) {
+            Files.deleteIfExists(pathToWorkingCopy);
+        }
     }
 
     @Test
@@ -43,17 +50,15 @@ public class CityCrudRepoTest {
 
     @Test
     void findByIdPositive() {
-        Integer cityId = 10504604;
-        Optional<City> city = cityIntegerCrudRepo.findById(cityId);
+        Optional<City> city = cityIntegerCrudRepo.findById(423);
         City result = city.get();
 
-        assertEquals(result.getId(), cityId);
+        assertEquals(result.getId(), 423);
     }
 
     @Test
     void findByIdNegative() {
-        Integer cityId = 10504604;
-        Optional<City> city = cityIntegerCrudRepo.findById(cityId);
+        Optional<City> city = cityIntegerCrudRepo.findById(423);
         City result = city.get();
 
         assertNotEquals(result.getId(), 33231);
@@ -62,7 +67,7 @@ public class CityCrudRepoTest {
     @Test
     void findByField() {
         String field = "name";
-        String value = "Байконур";
+        String value = "Dunwich";
         Optional<List<City>> cities = cityIntegerCrudRepo.findByField(field, value);
 
         assertEquals(2, cities.get().size());
@@ -70,47 +75,42 @@ public class CityCrudRepoTest {
 
     @Test
     void addCity() {
-        City city = new City();
         int id = ThreadLocalRandom.current().nextInt(Integer.MAX_VALUE);
-        city.setId(id);
 
-        city.setCountry_id(23);
-        city.setRegion_id(123);
-        city.setName("Impl");
-
-        cityIntegerCrudRepo.create(city);
+        cityIntegerCrudRepo.create(createCity(id, 10, 10, "Malcolm"));
 
         Optional<City> expOpt = cityIntegerCrudRepo.findById(id);
+        City actual =  cityIntegerCrudRepo.findById(id).get();
         City expected = expOpt.get();
 
-        assertEquals(expected.getName(), city.getName());
+        assertEquals(expected.getName(), actual.getName());
     }
 
     @Test
     void updateCity() {
-        Integer cityId = 4321;
-
-        City city = new City();
-        city.setId(cityId);
-        city.setCountry_id(23);
-        city.setRegion_id(123);
-        city.setName("Impl");
-
+        City city = cityIntegerCrudRepo.findById(423).get();
+        city.setName("San-Diego");
         cityIntegerCrudRepo.update(city);
 
-        Optional<City> expOpt = cityIntegerCrudRepo.findById(cityId);
-        City expected = expOpt.get();
-
-        assertTrue(expected.getName().equals("Impl"));
+        assertTrue(cityIntegerCrudRepo.findById(423).get().getName().equals("San-Diego"));
     }
 
     @Test
     void deleteCity() {
-        Integer cityId = 4326;
+        Integer cityId = 777;
 
         assertTrue(cityIntegerCrudRepo.findById(cityId).isPresent());
         cityIntegerCrudRepo.delete(cityId);
 
         assertTrue(cityIntegerCrudRepo.findById(cityId).isEmpty());
+    }
+
+    private static City createCity(int id, int regionId, int countryId, String name) {
+        City city = new City();
+        city.setId(id);
+        city.setRegion_id(regionId);
+        city.setCountry_id(countryId);
+        city.setName(name);
+        return city;
     }
 }
